move selector tests to integration tests
